+++ tcc1/src/main/java/com/tccagil/tcc1/domain/tarefas/AtividadesRecord.java
@@ -1,5 +1,5 @@
 package com.tccagil.tcc1.domain.tarefas;
 
-public record AtividadesRecord(String descricaoAtividade, int tarefaIdAtividade, int usuarioIdAtividade) {
+public record AtividadesRecord(String descricaoAtividade, Long tarefaIdAtividade, int usuarioIdAtividade) {
     
 }
